requirements.txt
autogen-agentchat
autogen-ext[semantic-kernel-google]
python-pptx==0.6.23
pytest
pytest-mock
pyyaml
python-dotenv
piper-tts
onnxruntime-gpu
whisperx
-e ./third_party/Orpheus-TTS/orpheus_tts_pypi
jiwer
cmudict
pronouncing
rapidfuzz
nltk
numpy==1.23.4
scipy==1.10.1
pybind11==2.11.1
git+https://github.com/aliutkus/speechmetrics@master#egg=speechmetrics[gpu]
soundfile
librosa
torch
faiss-cpu
